Make ols mock server available to all test classes
